chore: create useSelectTriggerTags
